incorporated more material UI components to forms
